# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Hikāya API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface AuthUser {
  id: string;
  /** @nullable */
  email: string | null;
  /** @nullable */
  firstName: string | null;
  /** @nullable */
  lastName: string | null;
  /** @nullable */
  profileImageUrl: string | null;
  role: string;
  preferredLanguage: string;
}

export interface AuthUserEnvelope {
  user: AuthUser | null;
}

export interface MobileTokenExchangeRequest {
  /** @minLength 1 */
  code: string;
  /** @minLength 1 */
  code_verifier: string;
  /** @minLength 1 */
  redirect_uri: string;
  /** @minLength 1 */
  state: string;
  /** @minLength 1 */
  nonce?: string;
}

export interface MobileTokenExchangeSuccess {
  token: string;
}

export const LogoutSuccessValue = {
  success: true,
} as const;
export type LogoutSuccess = typeof LogoutSuccessValue;

export interface ErrorEnvelope {
  error: string;
}

export interface UpdatePreferencesBody {
  preferredLanguage: string;
}

export interface Story {
  id: string;
  slug: string;
  title: string;
  description: string;
  /** @nullable */
  coverImage?: string | null;
  language: string;
  category: string;
  type: string;
  tags: string[];
  status: string;
  /** @nullable */
  videoUrl?: string | null;
  accentColor: string;
  createdAt: string;
  chapterCount?: number;
}

export interface Character {
  id: string;
  storyId: string;
  name: string;
  role: string;
  voiceId: string;
  tone: string;
  color: string;
  stability: number;
  similarity: number;
  style: number;
}

export type ScriptSegmentKind =
  (typeof ScriptSegmentKind)[keyof typeof ScriptSegmentKind];

export const ScriptSegmentKind = {
  narrator: "narrator",
  dialogue: "dialogue",
  scene: "scene",
} as const;

export interface ScriptSegment {
  index: number;
  kind: ScriptSegmentKind;
  /** @nullable */
  characterId?: string | null;
  characterName: string;
  emotion: string;
  text: string;
  /** @nullable */
  audioUrl?: string | null;
}

export interface Chapter {
  id: string;
  storyId: string;
  chapterNumber: number;
  title: string;
  content: string;
  segments: ScriptSegment[];
  durationSec: number;
  hasAudio?: boolean;
}

export interface ChapterSummary {
  id: string;
  chapterNumber: number;
  title: string;
  durationSec: number;
  hasAudio: boolean;
}

export interface ProgressEntry {
  storyId: string;
  chapterNumber: number;
  progressPercent: number;
  updatedAt: string;
  story: Story;
}

export interface StoryDetail {
  story: Story;
  characters: Character[];
  chapters: ChapterSummary[];
  bookmarked?: boolean;
  progress?: ProgressEntry | null;
}

export interface ChapterDetail {
  story: Story;
  chapter: Chapter;
  characters: Character[];
  /** @nullable */
  nextChapterNumber?: number | null;
  /** @nullable */
  prevChapterNumber?: number | null;
}

export interface UpsertProgressBody {
  storyId: string;
  chapterNumber: number;
  progressPercent: number;
}

export interface ToggleBookmarkBody {
  storyId: string;
}

export interface ToggleBookmarkResult {
  bookmarked: boolean;
}

export interface VoiceOption {
  voiceId: string;
  name: string;
  /** @nullable */
  category?: string | null;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  previewUrl?: string | null;
  /** @nullable */
  gender?: string | null;
  /** @nullable */
  language?: string | null;
}

export interface CreateStoryBody {
  title: string;
  slug: string;
  description?: string;
  /** @nullable */
  coverImage?: string | null;
  language: string;
  category: string;
  type: string;
  tags?: string[];
  /** @nullable */
  videoUrl?: string | null;
  accentColor?: string;
}

export interface UpdateStoryBody {
  title?: string;
  slug?: string;
  description?: string;
  /** @nullable */
  coverImage?: string | null;
  language?: string;
  category?: string;
  type?: string;
  tags?: string[];
  status?: string;
  /** @nullable */
  videoUrl?: string | null;
  accentColor?: string;
}

export interface CreateCharacterBody {
  name: string;
  role?: string;
  voiceId: string;
  tone?: string;
  color?: string;
  stability?: number;
  similarity?: number;
  style?: number;
}

export interface CreateChapterBody {
  chapterNumber: number;
  title: string;
  content?: string;
  segments?: ScriptSegment[];
}

export interface UpdateChapterBody {
  title?: string;
  content?: string;
  segments?: ScriptSegment[];
}

export interface GenerateAudioResult {
  generatedSegments: number;
  totalSegments: number;
  message?: string;
}

export interface ParsePdfBody {
  base64: string;
  filename?: string;
}

export interface ParsePdfResult {
  text: string;
  pages: number;
}

export interface AdminStats {
  totalStories: number;
  totalUsers: number;
  totalChapters: number;
  audioSegments: number;
}

/**
 * Opaque session token — `Bearer <sid>`.
 */
export type AuthorizationSessionHeaderParameter = string;

export type BeginBrowserLoginParams = {
  returnTo?: string;
};

export type HandleBrowserLoginCallbackParams = {
  code?: string;
  state?: string;
  iss?: string;
};

export type ListStoriesParams = {
  language?: string;
  category?: string;
  type?: string;
  q?: string;
};