# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Hikāya API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/stories:
    get:
      tags: [admin]
      operationId: adminListStories
      summary: List all stories (admin)
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Story'
    post:
      tags: [admin]
      operationId: createStory
      summary: Create story
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateStoryBody'
      responses:
        '201':
          description: created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Story'
  /admin/stories/{id}:
    patch:
      tags: [admin]
      operationId: updateStory
      summary: Update story
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateStoryBody'
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Story'
    delete:
      tags: [admin]
      operationId: deleteStory
      summary: Delete story
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '204':
          description: deleted
  /admin/stories/{id}/characters:
    post:
      tags: [admin]
      operationId: createCharacter
      summary: Add character to story
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCharacterBody'
      responses:
        '201':
          description: created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Character'
  /admin/characters/{id}:
    delete:
      tags: [admin]
      operationId: deleteCharacter
      summary: Delete character
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '204':
          description: deleted
  /admin/stories/{id}/chapters:
    post:
      tags: [admin]
      operationId: createChapter
      summary: Add chapter to story
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateChapterBody'
      responses:
        '201':
          description: created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Chapter'
  /admin/chapters/{id}:
    patch:
      tags: [admin]
      operationId: updateChapter
      summary: Update chapter (content, segments, title)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateChapterBody'
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Chapter'
    delete:
      tags: [admin]
      operationId: deleteChapter
      summary: Delete chapter
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '204':
          description: deleted
  /admin/chapters/{id}/generate-audio:
    post:
      tags: [admin]
      operationId: generateChapterAudio
      summary: Generate multi-voice audio for all segments in a chapter
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenerateAudioResult'
  /admin/parse-pdf:
    post:
      tags: [admin]
      operationId: parsePdf
      summary: Extract text from a PDF (base64 encoded)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ParsePdfBody'
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ParsePdfResult'
...
components:
...
  schemas:
...
    Story:
      type: object
      required: [id, slug, title, description, language, category, type, tags, status, accentColor, createdAt]
      properties:
        id: { type: string }
        slug: { type: string }
        title: { type: string }
        description: { type: string }
        coverImage: { type: [string, 'null'] }
        language: { type: string }
        category: { type: string }
        type: { type: string }
        tags:
          type: array
          items: { type: string }
        status: { type: string }
        videoUrl: { type: [string, 'null'] }
        accentColor: { type: string }
        createdAt: { type: string, format: date-time }
        chapterCount: { type: integer }
    Character:
      type: object
      required: [id, storyId, name, role, voiceId, tone, color, stability, similarity, style]
      properties:
        id: { type: string }
        storyId: { type: string }
        name: { type: string }
        role: { type: string }
        voiceId: { type: string }
        tone: { type: string }
        color: { type: string }
        stability: { type: integer }
        similarity: { type: integer }
        style: { type: integer }
    Chapter:
      type: object
      required: [id, storyId, chapterNumber, title, content, segments, durationSec]
      properties:
        id: { type: string }
        storyId: { type: string }
        chapterNumber: { type: integer }
        title: { type: string }
        content: { type: string }
        segments:
          type: array
          items: { $ref: '#/components/schemas/ScriptSegment' }
        durationSec: { type: integer }
        hasAudio: { type: boolean }
...
    ScriptSegment:
      type: object
      required: [index, kind, characterName, emotion, text]
      properties:
        index: { type: integer }
        kind:
          type: string
          enum: [narrator, dialogue, scene]
        characterId: { type: [string, 'null'] }
        characterName: { type: string }
        emotion: { type: string }
        text: { type: string }
        audioUrl: { type: [string, 'null'] }
...
    CreateStoryBody:
      type: object
      required: [title, slug, language, category, type]
      properties:
        title: { type: string }
        slug: { type: string }
        description: { type: string }
        coverImage: { type: [string, 'null'] }
        language: { type: string }
        category: { type: string }
        type: { type: string }
        tags:
          type: array
          items: { type: string }
        videoUrl: { type: [string, 'null'] }
        accentColor: { type: string }
    UpdateStoryBody:
      type: object
      properties:
        title: { type: string }
        slug: { type: string }
        description: { type: string }
        coverImage: { type: [string, 'null'] }
        language: { type: string }
        category: { type: string }
        type: { type: string }
        tags:
          type: array
          items: { type: string }
        status: { type: string }
        videoUrl: { type: [string, 'null'] }
        accentColor: { type: string }
    CreateCharacterBody:
      type: object
      required: [name, voiceId]
      properties:
        name: { type: string }
        role: { type: string }
        voiceId: { type: string }
        tone: { type: string }
        color: { type: string }
        stability: { type: integer }
        similarity: { type: integer }
        style: { type: integer }
    CreateChapterBody:
      type: object
      required: [chapterNumber, title]
      properties:
        chapterNumber: { type: integer }
        title: { type: string }
        content: { type: string }
        segments:
          type: array
          items: { $ref: '#/components/schemas/ScriptSegment' }
    UpdateChapterBody:
      type: object
      properties:
        title: { type: string }
        content: { type: string }
        segments:
          type: array
          items: { $ref: '#/components/schemas/ScriptSegment' }
    GenerateAudioResult:
      type: object
      required: [generatedSegments, totalSegments]
      properties:
        generatedSegments: { type: integer }
        totalSegments: { type: integer }
        message: { type: string }
    ParsePdfBody:
      type: object
      required: [base64]
      properties:
        base64: { type: string }
        filename: { type: string }
    ParsePdfResult:
      type: object
      required: [text, pages]
      properties:
        text: { type: string }
        pages: { type: integer }
...