# lib/api-zod/src/generated/types/chapter.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Hikāya API
 * OpenAPI spec version: 0.1.0
 */
import type { ScriptSegment } from "./scriptSegment";

export interface Chapter {
  id: string;
  storyId: string;
  chapterNumber: number;
  title: string;
  content: string;
  segments: ScriptSegment[];
  durationSec: number;
  hasAudio?: boolean;
}