# lib/api-zod/src/generated/types/createChapterBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Hikāya API
 * OpenAPI spec version: 0.1.0
 */
import type { ScriptSegment } from "./scriptSegment";

export interface CreateChapterBody {
  chapterNumber: number;
  title: string;
  content?: string;
  segments?: ScriptSegment[];
}