# lib/api-zod/src/generated/types/createCharacterBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Hikāya API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateCharacterBody {
  name: string;
  role?: string;
  voiceId: string;
  tone?: string;
  color?: string;
  stability?: number;
  similarity?: number;
  style?: number;
}