# lib/api-zod/src/generated/types/story.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Hikāya API
 * OpenAPI spec version: 0.1.0
 */

export interface Story {
  id: string;
  slug: string;
  title: string;
  description: string;
  /** @nullable */
  coverImage?: string | null;
  language: string;
  category: string;
  type: string;
  tags: string[];
  status: string;
  /** @nullable */
  videoUrl?: string | null;
  accentColor: string;
  createdAt: Date;
  chapterCount?: number;
}